requirements.txt
flask
tensorflow
tensorflow-hub
numpy
pillow
gunicorn